Move swish object to separate file
